move some naming helpers to apihelpers
