+++ pkg/apps/util/util.go
@@ -20,9 +20,9 @@ import (
 	kcoreclient "k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset/typed/core/internalversion"
 	kdeplutil "k8s.io/kubernetes/pkg/controller/deployment/util"
 
+	"github.com/openshift/origin/pkg/api/apihelpers"
 	deployapi "github.com/openshift/origin/pkg/apps/apis/apps"
 	deployapiv1 "github.com/openshift/origin/pkg/apps/apis/apps/v1"
-	"github.com/openshift/origin/pkg/util/namer"
 )
 
 var (
@@ -143,7 +143,7 @@ const DeployerPodSuffix = "deploy"
 
 // DeployerPodNameForDeployment returns the name of a pod for a given deployment
 func DeployerPodNameForDeployment(deployment string) string {
-	return namer.GetPodName(deployment, DeployerPodSuffix)
+	return apihelpers.GetPodName(deployment, DeployerPodSuffix)
 }
 
 // LabelForDeployment builds a string identifier for a Deployment.
