+++ pkg/serviceaccounts/util/helpers.go
@@ -4,7 +4,7 @@ import (
 	kapi "k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/v1"
 
-	"github.com/openshift/origin/pkg/util/namer"
+	"github.com/openshift/origin/pkg/api/apihelpers"
 )
 
 const (
@@ -15,7 +15,7 @@ const (
 )
 
 func GetDockercfgSecretNamePrefix(serviceAccount *kapi.ServiceAccount) string {
-	return namer.GetName(serviceAccount.Name, "dockercfg-", maxSecretPrefixNameLength)
+	return apihelpers.GetName(serviceAccount.Name, "dockercfg-", maxSecretPrefixNameLength)
 }
 
 // GetTokenSecretNamePrefix creates the prefix used for the generated SA token secret. This is compatible with kube up until
@@ -23,11 +23,11 @@ func GetDockercfgSecretNamePrefix(serviceAccount *kapi.ServiceAccount) string {
 // string.
 // TODO fix the upstream implementation to be more like this.
 func GetTokenSecretNamePrefix(serviceAccount *kapi.ServiceAccount) string {
-	return namer.GetName(serviceAccount.Name, "token-", maxSecretPrefixNameLength)
+	return apihelpers.GetName(serviceAccount.Name, "token-", maxSecretPrefixNameLength)
 }
 
 func GetDockercfgSecretNamePrefixV1(serviceAccount *v1.ServiceAccount) string {
-	return namer.GetName(serviceAccount.Name, "dockercfg-", maxSecretPrefixNameLength)
+	return apihelpers.GetName(serviceAccount.Name, "dockercfg-", maxSecretPrefixNameLength)
 }
 
 // GetTokenSecretNamePrefix creates the prefix used for the generated SA token secret. This is compatible with kube up until
@@ -35,5 +35,5 @@ func GetDockercfgSecretNamePrefixV1(serviceAccount *v1.ServiceAccount) string {
 // string.
 // TODO fix the upstream implementation to be more like this.
 func GetTokenSecretNamePrefixV1(serviceAccount *v1.ServiceAccount) string {
-	return namer.GetName(serviceAccount.Name, "token-", maxSecretPrefixNameLength)
+	return apihelpers.GetName(serviceAccount.Name, "token-", maxSecretPrefixNameLength)
 }
