--- pkg/util/namer/doc.go
@@ -1,2 +1,0 @@
-// Package namer contains a name generator for unique names
-package namer
